requirements.txt
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
Pillow==10.2.0
opencv-python-headless==4.9.0.80
plotly==5.20.0
requests==2.31.0
librosa==0.10.1
streamlit==1.32.2
st-theme==1.2.2
tensorflow==2.16.1